+++ tests/test_toml_document.py
@@ -192,6 +192,25 @@ name = "Test 1"
     assert doc["foo"]["bar"]["tests"][0]["name"] == "Test 1"
 
 
+def test_document_with_new_sub_table_after_other_other_table():
+    content = """[foo]
+name = "Bar"
+
+[bar]
+name = "Baz"
+
+[foo.baz]
+name = "Test 1"
+"""
+
+    doc = parse(content)
+    assert doc["foo"]["name"] == "Bar"
+    assert doc["bar"]["name"] == "Baz"
+    assert doc["foo"]["baz"]["name"] == "Test 1"
+
+    assert doc.as_string() == content
+
+
 def test_inserting_after_element_with_no_new_line_adds_a_new_line():
     doc = parse("foo = 10")
     doc["bar"] = 11