+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Fixed comma handling when parsing arrays. (Thanks to [@njalerikson](https://github.com/njalerikson))
 - Fixed comma handling when parsing inline tables. (Thanks to [@njalerikson](https://github.com/njalerikson))
+- Fixed a `KeyAlreadyPresent` error when declaring a sub table after other tables.
 
 
 ## [0.4.6] - 2018-10-16